+++ src/main/java/ru/m2mcom/natmob/domain/Hot_plays.java
@@ -61,7 +61,7 @@ public class Hot_plays implements Serializable {
     private Boolean checked;
 
     @Column(name = "rank_id")
-    private Integer rank_id;
+    private Long rank_id;
 
     @NotNull
     @Column(name = "sum_b", nullable = false)
@@ -212,16 +212,16 @@ public class Hot_plays implements Serializable {
         this.checked = checked;
     }
 
-    public Integer getRank_id() {
+    public Long getRank_id() {
         return rank_id;
     }
 
-    public Hot_plays rank_id(Integer rank_id) {
+    public Hot_plays rank_id(Long rank_id) {
         this.rank_id = rank_id;
         return this;
     }
 
-    public void setRank_id(Integer rank_id) {
+    public void setRank_id(Long rank_id) {
         this.rank_id = rank_id;
     }
